# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AAELink HTTP API",
    "version": "0.0.41-alpha",
    "summary": "Auto-generated route inventory. Method/path surface only — request/response schemas not yet introspected."
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Replace with your AAELink deployment URL"
    }
  ],
...
  "paths": {
...
    "/api/messages/scheduled": {
      "get": {
        "summary": "GET /api/messages/scheduled",
        "tags": [
          "messages"
        ],
        "description": "_marker:_ slack-compat surface (intentionally addressable, may be invoked by Slack-shaped clients)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/messages/scheduled",
        "tags": [
          "messages"
        ],
        "description": "_marker:_ slack-compat surface (intentionally addressable, may be invoked by Slack-shaped clients)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "DELETE /api/messages/scheduled",
        "tags": [
          "messages"
        ],
        "description": "_marker:_ slack-compat surface (intentionally addressable, may be invoked by Slack-shaped clients)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/notifications/email": {
      "get": {
        "summary": "GET /api/notifications/email",
        "tags": [
          "notifications"
        ],
        "description": "_marker:_ external integration entry point (webhook / IdP / push provider / device)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/notifications/email",
        "tags": [
          "notifications"
        ],
        "description": "_marker:_ external integration entry point (webhook / IdP / push provider / device)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}